Add S3 full tasks

+++ entrypoint.sh
@@ -30,24 +30,35 @@ fi
 # Create a dedicated profile for this action to avoid conflicts
 # with past/future actions.
 # https://github.com/jakejarvis/s3-rm-action/issues/1
-aws configure --profile s3-rm-action <<-EOF > /dev/null 2>&1
+aws configure --profile s3-action <<-EOF > /dev/null 2>&1
 ${AWS_ACCESS_KEY_ID}
 ${AWS_SECRET_ACCESS_KEY}
 ${AWS_REGION}
 text
 EOF
 
-# Sync using our dedicated profile and suppress verbose messages.
+if [ -z "$IS_DELETE" ]; then
+# Delete using our dedicated profile and suppress verbose messages.
 # All other flags are optional via the `args:` directive.
 sh -c "aws s3 rm s3://${AWS_S3_BUCKET}/${PATH_TO_DELETE} \
-              --profile s3-rm-action \
+              --profile s3-action \
               ${ENDPOINT_APPEND} $*"
+fi              
+
+if [ -z "$IS_SYNC" ]; then
+# Sync using our dedicated profile and suppress verbose messages.
+# All other flags are optional via the `args:` directive.
+sh -c "aws s3 sync ${SOURCE_DIR:-.} s3://${AWS_S3_BUCKET}/${DEST_DIR} \
+              --profile s3-action \
+              --no-progress \
+              ${ENDPOINT_APPEND} $*"
+fi
 
 # Clear out credentials after we're done.
 # We need to re-run `aws configure` with bogus input instead of
 # deleting ~/.aws in case there are other credentials living there.
 # https://forums.aws.amazon.com/thread.jspa?threadID=148833
-aws configure --profile s3-rm-action <<-EOF > /dev/null 2>&1
+aws configure --profile s3-action <<-EOF > /dev/null 2>&1
 null
 null
 null
